refactor: clean up utils - icons
